+++ iSistemium/Classes/Views/STMSupplyOrders/STMSupplyOrderOperationsTVC.m
@@ -773,7 +773,7 @@
     
     if (self.isInActiveTab) {
         
-        NSLog(@"barCodeScanner receiveBarCode: %@ withType:%d", barcode, type);
+        NSLog(@"barCodeScanner receiveBarCode: %@ withType:%lu", barcode, (unsigned long)type);
         
         switch (type) {
             case STMBarCodeTypeUnknown: {
